+++ src/main/java/com/team8/kanban/domain/board/controller/BoardController.java
@@ -5,6 +5,7 @@ import com.team8.kanban.domain.board.dto.BoardRequestDto;
 import com.team8.kanban.domain.board.dto.BoardResponseDto;
 import com.team8.kanban.domain.board.dto.BoardUserResponseDto;
 import com.team8.kanban.domain.board.service.BoardService;
+import com.team8.kanban.domain.board.service.BoardServiceImpl;
 import com.team8.kanban.global.common.CommonResponse;
 import com.team8.kanban.global.security.UserDetailsImpl;
 import java.util.List;
